11/23/2017

+++ AionHR.Web.UI.Forms/Login.aspx.cs
@@ -118,7 +118,7 @@ namespace AionHR.Web.UI.Forms
 
                 _systemService.SessionHelper.Set("CompanyName", getACResponse.result.companyName);
 
-                _systemService.SessionHelper.SetUserType((int)response.User.userType);
+                _systemService.SessionHelper.SetUserType(response.User.userType);
                 _systemService.SessionHelper.Set("CurrentUserName", userName);
 
                 _systemService.SessionHelper.Set("IsAdmin", response.User.isAdmin);
